+++ foundationdb-sys/Cargo.toml
@@ -20,9 +20,14 @@ license = "MIT/Apache-2.0"
 travis-ci = { repository = "bluejekyll/foundationdb-rs" }
 # codecov = { repository = "bluejekyll/foundationdb-rs", branch = "master", service = "github" }
 
+[features]
+default = [ "fdb-5_1" ]
+
+fdb-5_1 = [ ]
+fdb-5_2 = [ ]
+fdb-6_0 = [ ]
 
 [dependencies]
-
 
 [build-dependencies]
 bindgen = "0.36"
