+++ foundationdb-sys/src/lib.rs
@@ -2,14 +2,9 @@
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 
-include!("bindings.rs");
+include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_api_version() {
-        assert_eq!(FDB_API_VERSION, 510);
-    }
 }
