--- foundationdb-sys/wrapper.h
@@ -1,3 +1,0 @@
-#define FDB_API_VERSION 510
-
-#include <fdb_c.h>
